Fix potential race condition leading to panic
gas_price_with_cache first acquires the mutex and then compares the
existing timestamp to the passed in now. Since the `now` comes from
before the lock it is possible that the cache already got updated by a
simultaneous call in which case the cached timestamp can be larger than
now.
This is fixed by using checked_duration_since instead of
duration_since. The former does not panic. If the race condition happens
we instead correctly detect that we should use the cached entry since it
is new than now.